create custom state history animation functions and update playUpdatedStateHistory to use them
